update gitignore; separate game logic into its own class
