+++ server/engine/handlers/.gitignore
@@ -1,0 +1,8 @@
+swap files
+*swp
+
+compiled python files
+*.pyc
+
+env files
+.env
