+++ server/engine/handlers/games/basic_game.py
@@ -1,28 +1,41 @@
-import torch
-from copy import deepcopy
+import numpy as np, pdb
 
-from games.utils import is_game_finished, calc_reward, process_action
+def _init_action_set():
+    move_up = lambda x, y: ( x, y - 1 )
+    move_down = lambda x, y: ( x, y + 1 )
+    move_left = lambda x, y: ( x - 1, y )
+    move_right = lambda x, y: ( x + 1, y )
+    
+    return [ move_up, move_down, move_left, move_right ]
 
-class GameEnv:
-    def __init__(self, height, width, obstacles):
-        self.height = height
-        self.width = width
-        self.obstacles = obstacles
-        self.grid = None
-        self.dtype = torch.FloatTensor
+class BasicGame:
+    def __init__(self):
+        self.action_set = _init_action_set()
 
-    def reset(self):
-        grid = torch.zeros((1, 1, self.height, self.width)).type(self.dtype)
-        #  mark the player starting positions
-        grid[:, :, self.height - 1, 0] = 1
-        grid[:, :, 0, self.width - 1] = 2
-        for obstacle_pos in self.obstacles:
-            i, j = obstacle_pos
-            grid[:, :, i, j] = -1
-        self.grid = grid
+    def is_game_finished(self, grid):
+        _, _, height, width = grid.size()
+        return grid.numpy()[:, : 0, height - 1] == 2 \
+                or grid.numpy()[:, :, width-1, 0] == 1 
 
-    def step(self, action, turn):
-        new_state = process_action(deepcopy(self.grid), action, turn)
-        reward = self.dtype([ calc_reward(new_state) ])
-        done = is_game_finished(new_state)
-        return new_state, reward, done, None
+    def calc_reward(self, grid):
+        return 0
+
+    def process_action(self, grid, action, turn):
+        _, _, height, width = grid.size()
+        state = grid.numpy()
+
+        #  determine the new position of player based on turn and action
+        _, _, y_old, x_old = np.where( state == ( turn + 1 ) )
+        x_new, y_new = self.action_set[action](x_old, y_old)
+
+
+        #  make sure the new position is valid (wrt game rules)
+        if not (0 < x_new <= width and 0 < y_new <= height):
+            return grid
+        if grid.numpy()[:, :, y_new, x_new] != 0:
+            return grid
+
+        #  update the grid with the player's new position
+        grid[:, :, x_old, y_old] = 0
+        grid[:, :, x_new, y_new] = turn + 1
+        return grid
